// features/api/model/supplierResource.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { LiftingResource } from './liftingResource';

export interface SupplierResource {
  id: string;
  name: string;
  current_balance: number;
  /** @nullable */
  email: string | null;
  /** @nullable */
  phone: string | null;
  liftings?: LiftingResource[];
  created_at: string;
  updated_at: string;
}