// features/api/model/creditorsIndex200.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { CreditorsIndex200Links } from './creditorsIndex200Links';
import type { CreditorsIndex200Meta } from './creditorsIndex200Meta';
import type { SupplierResource } from './supplierResource';

export type CreditorsIndex200 = {
  data: SupplierResource[];
  links: CreditorsIndex200Links;
  meta: CreditorsIndex200Meta;
};